requirements.txt
beautifulsoup4==4.8.2
certifi==2019.9.11
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.2.6
django-admin-autoregister==1.0.0
django-cors-headers==3.1.1
django-model-utils==3.2.0
django-rest-auth==0.9.5
django-rest-swagger==2.2.0
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
idna==2.8
itypes==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
openapi-codec==1.3.2
Pillow==6.2.1
PyJWT==1.7.1
pytz==2019.3
PyYAML==5.1.2
requests==2.22.0
simplejson==3.16.0
six==1.12.0
soupsieve==2.0
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.25.6
